Part 1
Connected URL patterns for blog project and the App.
For index.html added it to a root template folder and set a path for access in the settings.py
creaded a model for blog attribute.
Added dummy contents for each blog posts and loop through it to show it in the index.html views